+++ src/pharmpy/tools/allometry/tool.py
@@ -10,10 +10,9 @@ from pharmpy.modeling import (
     add_allometry,
     get_pk_parameters,
     summarize_errors,
-    summarize_individuals,
-    summarize_individuals_count_table,
     summarize_modelfit_results,
 )
+from pharmpy.tools import summarize_individuals, summarize_individuals_count_table
 from pharmpy.tools.modelfit import create_fit_workflow
 from pharmpy.utils import runtime_type_check, same_arguments_as
 from pharmpy.workflows import Task, Workflow